+++ src/truflation/data/connectors/rest_to_csv.py
@@ -4,7 +4,7 @@ Rest To CSV connector
 
 import requests
 import pandas as pd
-from truflation.data.connector import Connector
+from truflation.data.connectors.base import Connector
 
 class RestToCsvConnector(Connector):
     def __init__(self, rest_url, csv_path):
